modify return value of get_time_sec to u64

+++ libtime/libtime.c
@@ -27,7 +27,8 @@
 
 #define TIME_FORMAT "%Y%m%d%H%M%S"
 
-uint32_t time_get_sec()
+// uint32_t time_get_sec()
+uint64_t time_get_sec()
 {
     time_t t;
     t = time(NULL);
